Fix expansion of text editor
